+++ src/main/java/superapp/logic/concreteServices/SuperAppObjectService.java
@@ -20,9 +20,9 @@ import superapp.util.exceptions.ForbbidenOperationException;
 import superapp.util.exceptions.InvalidInputException;
 import superapp.util.exceptions.NotFoundException;
 import superapp.util.EmailChecker;
+
 import static superapp.data.UserRole.*;
-import static superapp.util.ControllersConstants.DEFAULT_SORTING_DIRECTION;
-import static superapp.util.ControllersConstants.DEPRECATED_EXCEPTION;
+import static superapp.util.Constants.*;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -60,6 +60,11 @@ public class SuperAppObjectService extends AbstractService implements AdvancedSu
                 createdBy.getSuperapp().isEmpty() ||
                 !EmailChecker.isValidEmail(createdBy.getEmail()))
             throw new InvalidInputException("Invalid creating user details");
+
+        if (!this.isValidUserCredentials(new UserPK(createdBy.getSuperapp(),createdBy.getEmail()),
+                SUPERAPP_USER, this.userRepository))
+            throw new ForbbidenOperationException(SUPERAPP_USER_ONLY_EXCEPTION);
+
 
         Boolean active = object.getActive();
         if (active == null)
@@ -116,11 +121,11 @@ public class SuperAppObjectService extends AbstractService implements AdvancedSu
 
     @Override
     @Transactional
-    public SuperAppObjectBoundary updateObject(String objectSuperapp,
-                                               String internalObjectId,
-                                               SuperAppObjectBoundary update,String userSuperapp,String email) {
+    public SuperAppObjectBoundary updateObject(String objectSuperapp, String internalObjectId,
+                                               SuperAppObjectBoundary update, String userSuperapp, String email) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        if(!this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            throw new ForbbidenOperationException(SUPERAPP_USER_ONLY_EXCEPTION);
 
         Optional<SuperAppObjectEntity> objectO =
                 this.objectRepository.findById(new SuperappObjectPK(objectSuperapp, internalObjectId));
@@ -159,10 +164,10 @@ public class SuperAppObjectService extends AbstractService implements AdvancedSu
     @Override
     @Transactional
     public void bindNewChild(String parentSuperapp, String parentObjectId,
-                             SuperAppObjectIdBoundary newChild,
-                             String userSuperapp, String email) {
+                             SuperAppObjectIdBoundary newChild, String userSuperapp, String email) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        if(!this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            throw new ForbbidenOperationException(SUPERAPP_USER_ONLY_EXCEPTION);
 
         SuperAppObjectEntity parent = this.objectRepository
                 .findById(new SuperappObjectPK(parentSuperapp, parentObjectId))
@@ -180,16 +185,24 @@ public class SuperAppObjectService extends AbstractService implements AdvancedSu
 
     @Override
     @Transactional(readOnly = true)
-    public SuperAppObjectBoundary getSpecificObject(String objectSuperapp, String internalObjectId, String userSuperapp, String email) {
+    public SuperAppObjectBoundary getSpecificObject(String objectSuperapp, String internalObjectId,
+                                                    String userSuperapp, String email) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        Optional<SuperAppObjectEntity> objectE = this.objectRepository.findById(new SuperappObjectPK(objectSuperapp, internalObjectId));
 
-        Optional<SuperAppObjectEntity> objectE =
-                this.objectRepository.findById(new SuperappObjectPK(objectSuperapp, internalObjectId));
         if (objectE.isEmpty())
             throw new NotFoundException("Object does not exist");
 
-        return this.converter.toBoundary(objectE.get());
+        if (this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            return this.converter.toBoundary(objectE.get());
+
+        else if (this.isValidUserCredentials(userId, MINIAPP_USER, this.userRepository)) {
+             if (!objectE.get().getActive())
+                throw new NotFoundException("Requested inactive object");
+
+             return this.converter.toBoundary(objectE.get());
+     }
+        throw new ForbbidenOperationException(SUPERAPP_MINIAPP_USERS_ONLY_EXCEPTION);
     }
 
     @Override
@@ -198,15 +211,15 @@ public class SuperAppObjectService extends AbstractService implements AdvancedSu
                                                     String userSuperapp, String email,
                                                     int size, int page) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        PageRequest pageReq = PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "objectId");
 
-        return this.objectRepository.findAll(PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "objectId"))
-                .stream()
-                .filter(obj -> obj.getObjectId().equals(internalObjectId) && obj.getSuperapp().equals(objectSuperapp))
-                .map(SuperAppObjectEntity::getChildren)
-                .flatMap(superAppObjectEntities -> superAppObjectEntities.stream()
-                        .map(this.converter::toBoundary))
-                .collect(Collectors.toList());
+        if (this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            return getChildrenRepoSearch(pageReq, internalObjectId, userSuperapp ,true);
+
+        else if (this.isValidUserCredentials(userId, MINIAPP_USER, this.userRepository))
+            return getChildrenRepoSearch(pageReq,internalObjectId,userSuperapp,false);
+
+        throw new ForbbidenOperationException(SUPERAPP_MINIAPP_USERS_ONLY_EXCEPTION);
     }
 
     @Override
@@ -214,52 +227,60 @@ public class SuperAppObjectService extends AbstractService implements AdvancedSu
     public List<SuperAppObjectBoundary> getParents(String objectSuperapp, String internalObjectId,String userSuperapp,
                                                    String email, int size, int page) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        PageRequest pageReq = PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "objectId");
 
-        return this.objectRepository.findAll(PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "objectId"))
-                .stream()
-                .filter(obj -> obj.getObjectId().equals(internalObjectId) && obj.getSuperapp().equals(objectSuperapp))
-                .map(SuperAppObjectEntity::getParents)
-                .flatMap(superAppObjectEntities -> superAppObjectEntities.stream()
-                        .map(this.converter::toBoundary))
-                .collect(Collectors.toList());
+        if(this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            return getParentRepoSearch(pageReq, internalObjectId, objectSuperapp,true);
+
+        else if(this.isValidUserCredentials(userId, MINIAPP_USER, this.userRepository))
+            return getParentRepoSearch(pageReq, internalObjectId, objectSuperapp,false);
+
+        throw new ForbbidenOperationException(SUPERAPP_MINIAPP_USERS_ONLY_EXCEPTION);
+
     }
 
     @Override
     @Transactional(readOnly = true)
     public List<SuperAppObjectBoundary> getAllObjects(String userSuperapp, String email, int size, int page) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        PageRequest pageReq = PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "userEmail");
 
-        return this.objectRepository.findAll(PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION,
-                        "userSuperapp", "userEmail"))
-                    .stream()
-                    .map(this.converter::toBoundary)
-                    .collect(Collectors.toList());
+        if (this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            return findAllObjectsRepoSearch(pageReq,true);
+
+        else if(this.isValidUserCredentials(userId, MINIAPP_USER, this.userRepository))
+            return findAllObjectsRepoSearch(pageReq,false);
+
+        throw new ForbbidenOperationException(SUPERAPP_MINIAPP_USERS_ONLY_EXCEPTION);
     }
 
     @Override
     public List<SuperAppObjectBoundary> SearchObjectsByType(String type, String userSuperapp, String email, int size, int page) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        PageRequest pageReq = PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "objectId");
 
-        return this.objectRepository.findByType(type, PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "objectId"))
-                .stream()
-                .map(this.converter::toBoundary)
-                .collect(Collectors.toList());
+        if( isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            return findAllObjectsByTypeRepoSearch(pageReq, type,true);
+
+        else if (isValidUserCredentials(userId, MINIAPP_USER, this.userRepository))
+            return findAllObjectsByTypeRepoSearch(pageReq, type,false);
+
+        throw new ForbbidenOperationException(SUPERAPP_MINIAPP_USERS_ONLY_EXCEPTION);
     }
 
     @Override
     @Transactional
     public List<SuperAppObjectBoundary> SearchObjectsByExactAlias(String alias, String userSuperapp, String email, int size, int page) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        PageRequest pageReq = PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "objectId");
 
-        return this.objectRepository
-                .findByAlias(alias, PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "objectId"))
-                .stream()
-                .map(this.converter::toBoundary)
-                .collect(Collectors.toList());
+        if (this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            return findByAliasRepoSearch(pageReq,alias,true);
+
+        else if (this.isValidUserCredentials(userId, MINIAPP_USER, this.userRepository))
+            return findByAliasRepoSearch(pageReq,alias,false);
+
+        throw new ForbbidenOperationException(SUPERAPP_MINIAPP_USERS_ONLY_EXCEPTION);
     }
 
     @Override
@@ -267,20 +288,96 @@ public class SuperAppObjectService extends AbstractService implements AdvancedSu
     public List<SuperAppObjectBoundary> SearchObjectsByAliasContaining(String text, String userSuperapp, String email, int size, int page)
     {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository);
+        PageRequest pageReq = PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "superapp", "objectId");
 
-        return this.objectRepository
-                .findByAliasContaining(text, PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "objectId"))
-                .stream()
-                .map(this.converter::toBoundary)
-                .collect(Collectors.toList());
+        if (this.isValidUserCredentials(userId, SUPERAPP_USER, this.userRepository))
+            return findByAliasContainingRepoSearch(pageReq, text,true);
+
+        else if (this.isValidUserCredentials(userId, MINIAPP_USER, this.userRepository))
+            return findByAliasContainingRepoSearch(pageReq, text,false);
+
+        throw new ForbbidenOperationException(SUPERAPP_MINIAPP_USERS_ONLY_EXCEPTION);
+
     }
 
     @Override
     @Transactional
     public void deleteAllObjects(String userSuperapp, String email) {
         UserPK userId = new UserPK(userSuperapp, email);
-        this.isValidUserCredentials(userId, ADMIN, this.userRepository);
-        this.objectRepository.deleteAll();
+        if (isValidUserCredentials(userId, ADMIN, this.userRepository))
+            this.objectRepository.deleteAll();
+        else
+            throw new ForbbidenOperationException(SUPERAPP_USER_ONLY_EXCEPTION);
+    }
+
+    private List<SuperAppObjectBoundary> findAllObjectsRepoSearch(PageRequest pageReq, boolean isSuperAppUser) {
+        return this.objectRepository.findAll(pageReq)
+                .stream()
+                .map(this.converter::toBoundary)
+                .filter(object -> object.getActive()|| isSuperAppUser) // if MINIAPP_USER returns only active
+                .collect(Collectors.toList());
+    }
+
+    private List<SuperAppObjectBoundary> findByAliasRepoSearch(PageRequest pageReq, String alias, boolean isSuperAppUser){
+        return this.objectRepository.findByAlias(alias, pageReq)
+                .stream()
+                .map(this.converter::toBoundary)
+                .filter(object -> object.getActive() || isSuperAppUser) // if MINIAPP_USER returns only active
+                .collect(Collectors.toList());
+    }
+
+    private List<SuperAppObjectBoundary> findByAliasContainingRepoSearch(PageRequest pageReq, String text, boolean isSuperAppUser){
+        return this.objectRepository
+                .findByAliasContaining(text, pageReq)
+                .stream()
+                .filter(object -> object.getActive() || isSuperAppUser) // if MiniappUser get only active
+                .map(this.converter::toBoundary)
+                .collect(Collectors.toList());
+    }
+
+    private List<SuperAppObjectBoundary> getParentRepoSearch(PageRequest pageReq, String internalObjectId, String objectSuperapp, boolean isSuperAppUser) {
+        List<SuperAppObjectEntity> objectList =
+                this.objectRepository
+                .findAll(pageReq)
+                .stream()
+                .filter(obj -> obj.getObjectId().equals(internalObjectId) && obj.getSuperapp().equals(objectSuperapp))
+                .toList();
+
+        SuperAppObjectEntity requestedObject = objectList.isEmpty() ? null : objectList.get(0);
+        if (requestedObject == null || !(isSuperAppUser || requestedObject.getActive()))
+            return new ArrayList<SuperAppObjectBoundary>(0);
+
+        return requestedObject.getParents()
+                .stream()
+                .map(this.converter::toBoundary)
+                .filter(object -> object.getActive()|| isSuperAppUser)
+                .collect(Collectors.toList());
+    }
+
+    private List<SuperAppObjectBoundary> getChildrenRepoSearch(PageRequest pageReq, String internalObjectId, String objectSuperapp, boolean isSuperAppUser) {
+        List<SuperAppObjectEntity> objectList =
+                this.objectRepository
+                .findAll(pageReq)
+                .stream()
+                .filter(obj -> obj.getObjectId().equals(internalObjectId) && obj.getSuperapp().equals(objectSuperapp))
+                .toList();
+
+        SuperAppObjectEntity requestedObject = objectList.isEmpty() ? null : objectList.get(0);
+        if (requestedObject == null || !(isSuperAppUser || requestedObject.getActive()))
+            return new ArrayList<SuperAppObjectBoundary>(0);
+
+        return requestedObject.getChildren()
+                .stream()
+                .map(this.converter::toBoundary)
+                .filter(object -> object.getActive()|| isSuperAppUser)
+                .collect(Collectors.toList());
+    }
+
+    private List<SuperAppObjectBoundary> findAllObjectsByTypeRepoSearch(PageRequest pageReq, String type, boolean isSuperAppUser){
+        return this.objectRepository.findByType(type, pageReq)
+                .stream()
+                .map(this.converter::toBoundary)
+                .filter(object -> object.getActive()|| isSuperAppUser)// if MiniappUser get only active
+                .collect(Collectors.toList());
     }
 }
